Renaming folder structure and file names
