+++ applications/plugins/hid_app/assets/Ble_connected_15x15.png
