+++ applications/plugins/hid_app/assets/Ble_disconnected_15x15.png
