+++ applications/plugins/hid_app/assets/ButtonLeft_4x7.png
