+++ applications/plugins/hid_app/assets/ButtonUp_7x4.png
