+++ applications/plugins/hid_app/assets/Circles_47x47.png
