+++ applications/plugins/hid_app/assets/Left_mouse_icon_9x9.png
