+++ applications/plugins/hid_app/assets/Like_pressed_17x17.png
