+++ applications/plugins/hid_app/assets/Ok_btn_9x9.png
