+++ applications/plugins/hid_app/assets/Ok_btn_pressed_13x13.png
