+++ applications/plugins/hid_app/assets/Pin_arrow_left_9x7.png
