+++ applications/plugins/hid_app/assets/Pin_back_arrow_10x8.png
