+++ applications/plugins/hid_app/assets/Pressed_Button_13x13.png
