+++ applications/plugins/hid_app/assets/Right_mouse_icon_9x9.png
