+++ applications/plugins/hid_app/assets/Space_65x18.png
