+++ applications/plugins/hid_app/hid.c
@@ -1,5 +1,5 @@
-#include "bt_hid.h"
-#include "views.h"
+#include "hid.h"
+//#include "views.h"
 #include <furi_hal_bt.h>
 #include <furi_hal_usb.h>
 #include <notification/notification_messages.h>
