+++ applications/plugins/hid_app/hid_remote_10px.png
