+++ applications/plugins/hid_app/views.h
