+++ applications/plugins/hid_app/views/bt_hid_tiktok.c
@@ -4,7 +4,7 @@
 #include <furi_hal_usb_hid.h>
 #include <gui/elements.h>
 
-#include "bt_hid_icons.h"
+#include "hid_icons.h"
 
 #define TAG "BtHidApp"
