+++ applications/plugins/hid_app/views/hid_keynote.c
@@ -1,11 +1,11 @@
-#include "bt_hid_keynote.h"
+#include "hid_keynote.h"
 #include <furi.h>
 #include <furi_hal_bt_hid.h>
 #include <furi_hal_usb_hid.h>
 #include <gui/elements.h>
-#include "../bt_hid.h"
+#include "../hid.h"
 
-#include "bt_hid_icons.h"
+#include "hid_icons.h"
 
 #define TAG "HidApp"
