+++ applications/plugins/hid_app/views/hid_media.c
@@ -1,11 +1,11 @@
-#include "bt_hid_media.h"
+#include "hid_media.h"
 #include <furi.h>
 #include <furi_hal_bt_hid.h>
 #include <furi_hal_usb_hid.h>
 #include <gui/elements.h>
-#include "../bt_hid.h"
+#include "../hid.h"
 
-#include "bt_hid_icons.h"
+#include "hid_icons.h"
 
 #define TAG "HidApp"
