+++ applications/plugins/hid_app/views/hid_media.h
